fixing windows pipeline

+++ test/src/Barrier_test.cpp
@@ -746,7 +746,7 @@ TEST_P(BarrierTest, BarrierTest_timeout_race)
 
     auto count = GetParam();
 
-    tinycoro::Barrier barrier{std::max((count / 2), 1ul)};
+    tinycoro::Barrier barrier{std::max((count / (size_t)2), (size_t)1)};
 
     std::atomic<decltype(count)> cc = count;
